backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
python-docx>=1.1
anthropic>=0.39
pypdf>=4.2
